+++ sdk/identity/azure-identity/azure/identity/aio/_credentials/client_assertion.py
@@ -40,7 +40,19 @@ class ClientAssertionCredential(AsyncContextManager, GetTokenMixin):
 
     def __init__(self, tenant_id: str, client_id: str, func: Callable[[], str], **kwargs: Any) -> None:
         self._func = func
-        self._client = AadClient(tenant_id, client_id, **kwargs)
+        authority = kwargs.pop("authority", None)
+        cache = kwargs.pop("cache", None)
+        cae_cache = kwargs.pop("cae_cache", None)
+        additionally_allowed_tenants = kwargs.pop("additionally_allowed_tenants", None)
+        self._client = AadClient(
+            tenant_id,
+            client_id,
+            authority=authority,
+            cache=cache,
+            cae_cache=cae_cache,
+            additionally_allowed_tenants=additionally_allowed_tenants,
+            **kwargs
+        )
         super().__init__(**kwargs)
 
     async def __aenter__(self):